+++ Libraries/objectedGraphicsLibrary/texture.cpp
@@ -5,6 +5,8 @@
 //  Created by Jakob Zorz on 04/07/2020.
 //
 
+#include <utility>
+
 #include "singleWindowLibrary.hpp"
 #include "objectedGraphicsLibrary.hpp"
 
@@ -15,8 +17,8 @@ ogl::texture::texture(ogl::objectType type) {
 void ogl::texture::setTexture(SDL_Texture* input_texture, int width_, int height_) {
     freeTexture();
     texture_ = input_texture;
-    width = width_;
-    height = height_;
+    width = static_cast<unsigned short>(width_);
+    height = static_cast<unsigned short>(height_);
 }
 
 void ogl::texture::setTexture(SDL_Texture* input_texture) {
@@ -41,13 +43,13 @@ ogl::texture::~texture() {
 void ogl::texture::loadFromText(std::string text, SDL_Color text_color) {
     int temp_width, temp_height;
     setTexture(swl::loadTextureFromText(text, text_color, &temp_width, &temp_height));
-    width = temp_width;
-    height = temp_height;
+    width = static_cast<unsigned short>(temp_width);
+    height = static_cast<unsigned short>(temp_height);
 }
 
 void ogl::texture::loadFromFile(std::string path) {
     int temp_width, temp_height;
     setTexture(swl::loadTextureFromFile(path, &temp_width, &temp_height));
-    width = temp_width;
-    height = temp_height;
+    width = static_cast<unsigned short>(temp_width);
+    height = static_cast<unsigned short>(temp_height);
 }
